update package sources before performing `puppet` install
An issue can occur during the puppet install where certain packages are not found. This results in a 404 error which causes the puppet install to fail.

This ulimately results in the following output:
```
The `puppet` binary appears not to be in the PATH of the guest. This
could be because the PATH is not properly setup or perhaps Puppet is not
installed on this guest. Puppet provisioning can not continue without
Puppet properly installed.
```

Running `apt-get update` beforehand helps to mitigate these 404 issues by ensuring that the sources are up to date.