Fix wrong unison version downloaded
The Arch Linux unison package was recently updated to version 2.51. This change makes sure we still download version 2.48. As an aside, version 2.51 was compiled with version 2.28 of the libc6 package which is only available on the future Cosmic Ubuntu release. That's why I chose to keep the current unison version - updating to the newest would indeed requires too much effort for now.